move local_package content into fs
