+++ py-rattler/src/record.rs
@@ -1031,7 +1031,7 @@ impl PyRecord {
     fn from_package_archive(py: Python<'_>, path: PathBuf) -> PyResult<Bound<'_, PyAny>> {
         future_into_py(py, async move {
             let record =
-                rattler_package_streaming::local_package::repodata_record_from_package_archive(
+                rattler_package_streaming::fs::repodata_record_from_package_archive(
                     path,
                 )
                 .await
